Prevent errors after running /upgrade
After running /upgrade, all the buffers will still exist, but the script
is stopped and run again, so all the state in the script will be gone.
This causes a bunch of errors because buffers can't be created and the
script won't know about the existing buffers.

Instead of loading the script normally, print a note that wee-slack will
not work until it's reloaded. I didn't take the effort to make it work
automatically, because it should be easy for users just to reload it.

Fixes #275, fixes #309, fixes #310