Fix completion for /slack
Completion for our commands can't use || because they can be used as
subcommands for /slack and weechat doesn't support multiple templates
with the same subcommand. See https://github.com/weechat/weechat/issues/1481

This makes the completion for /reply slightly incorrect. It now allows
you to complete two thread arguments when not using -alsochannel. To fix
that, we would have to write a custom completion function. It's not
hard, but I'm not sure it's worth the effort for such a small detail.