Move store_message back into subprocess functions
In subprocess_thread_message we need to store the message before calling
notify_thread, so storing it after the function returned didn't work.

This is because when auto_open_threads is on and you get a new message,
it will open the thread and print the messages immediately, which will
fail because the message ts has been added to submessages, but the
message has not been stored yet.