Keep submessages from old message when storing new
Submessages (or replies) are not a part of the message anymore with
conversations.history, we have to call conversations.replies to get
them. So when fetching the channel history, the new message will have an
empty submessages. Keep the old, so we don't have to call
conversations.replies for each thread parent every time we call
conversations.history.