If message has blocks, only render blocks unless rich_text is present
Slack doesn't render the `text` property at all when `blocks` is
non-empty, so we shouldn't render both `text` and `blocks`. However,
because we don't support the `rich_text` block type yet, we have to fall
back to the `text` property when `rich_text` is present. Therefore,
render both `text` and `blocks` if `rich_text` is present, else render
only `blocks` if it is non-empty, and only `text` if `blocks` is empty.