Set last_read to SlackTS(0) when missing
When last_read is missing, we should assume no messages are read instead
of all up until the channel/message is created. The only time I know
this happens is for messages that are not yet thread parents.

This probably doesn't matter in practice though. For new thread parents
the messages that make it a thread parent will arrive after we create
the message, so they would have newer ts than the one we set for
last_read. Still, I think it makes more sense to set it to 0.