Use if/else instead of a dict in formatted_name
I mainly did this because I'm going to call formatted_name with
style=default before team is set, and that didn't work as it was because
the other styles used the team.

For SlackDMChannel I moved colorizing into the branch for the sidebar
style because colors should never be included in the other styles.
formatted_name is never called with enable_color=True and another style
than sidebar, so the behavior doesn't change.