Set last_read when (un)subscribing to a thread
When you subscribe to a thread, I think it's safe to assume that you
have read the messages or want to treat the messages as read, and this
is what the web client does.

I'm not sure what the point of sending last_read for unsubscribe is, but
the web client does so, and it was easiest to do it for both.