Fetch message replies if subscribed latest_reply > last_read
When using conversations.history (which doesn't include replies), we
want to fetch the replies for a thread even if
thread_messages_in_channel is off if the user is subscribed to the
thread and it has unread messages, so we can notify for new replies.

Technically, we could notify without fetching the history, but it's
easier to just fetch the history so the normal notification check is
run, and when the thread has unread messages, the user probably wants to
read them, so it's nice that they don't have to wait for history when
opening the thread.