Prevent multiple thread notifications when mentioned
There are two situations where notify_thread is called multiple times
where we only want to notify once. One is when we load history for a
thread which has multiple unread messages, the other is when you become
subscribed to a thread because you are mentioned in it.

For the first case, we want to set last_notify to now, so we don't
notify for the rest of the unread messages following the first. For the
second, the message ts may be greater than now (same second, but greater
minor part), so we want to set last_notify to the message ts. Therefore
we set last_notify to max of these two.