On load, fetch history after connected to websocket
The websocket url is only valid for 30 seconds. If wee-slack didn't
manage to connect in time, it would set history_needs_update and
reconnect, but not load history again which meant that each channel
would load history when you switched to the buffer.

I also think there was a slight chance of a message being sent in the
time between history loading and connecting to the websocket, which I
think would mean it would be lost. By fetching history after connecting
to the websocket, this can't happen.

Lastly, by not fetching history before connecting to the websocket, I
think there is less chance of not connecting to the websocket url before
it expires.