Set history_needs_update when receiving history
If wee-slack looses connection to the server while a history request is
in the queue and reconnects before the history response is received, it
will set history_needs_update to True and call get_history, but since a
history request is pending, a new one won't be triggered so
history_needs_update will remain True. Setting history_needs_update to
False when receiving the history fixes this.