When changing last message, print new lines if needed
WeeChat doesn't support inserting new lines apart from at the bottom, so
normally when a message is changed to contain more lines, it is not
shown properly until the buffer is redrawn. However, if the message
that's being changed is the last one, we can insert new lines so it's
shown properly.