Don't print short name prefix if set with /buffer set
When short name is set with /buffer set short_name, instead of changing
it to include the prefix, keep it as is. I think this is what would be
expected, and it also allows the user to choose which prefix to show.
When showing typing notices, replace the first character with the typing
character.