Support showing buflist presence when short_name is overridden
When you set a short name, the prefix wee-slack sets (# for channels,
+ for dms with present users etc.) is normally dropped so the short name
is exactly how the user wants it. However that makes it impossible to
see presence in the buflist when short name is set (unless you set it
with /label which keeps the prefix). Therefore, if the short name is set
to a name starting with a space, replace the space with + if the user is
present. This gives the same prefix behavior and characters as when
short name is not set.