Remove buffer notify option for unmuted buffers
We were setting the notify level for unmuted buffers to 3. This is not
the right thing to do because weechat has a
weechat.look.buffer_notify_default option which might be set to a
different value. By removing the option instead, weechat will use the
default value.

Additionally, when we change the notify level for a buffer, this is
stored in the config (for weechat >= 2.6). Having such a config for all
buffers is annoying, but with this change the config will only be
present for muted buffers.

This fix is not perfect because you still get config values for muted
buffers, and since wee-slack is overriding the notify level you can't
override it yourself for specific buffers. A potential solution to this
could be to set notify level for each line instead of setting notify
level for the buffer.

Fixes #746