Revert "Don't set no_highlight for backlog messages"
This caused a problem for relay clients. They are not notified when
print hooks are disabled/enabled, so they have to use the tags and line
properties to determine if a line should notify.

One could argue that they should not notify when notify_none is present,
even if it's a highlight, but this is not how the hotlist or default
beep trigger in weechat behaves. When a message with a highlight is
printed, that is added to the hotlist and beeps even if notify_none is
set (wee-slack removes the buffer from the hotlist immediately after
printing the line, which is why that wasn't a problem in plain weechat).

Additionally, this caused messages you had reacted to to be highlighted
when the buffer was reprinted, since the text now included your name
when the line was printed.

This reverts commit de5eb518fda26fed3b5cd41c20b064c04020bbb9.