Revert "Use buflist.format.indent for thread short name indent"
This caused some issues, so I think it's best to revert it:

 - If you don't set a color in the indent option, and otherwise use
   default format options, the normal indent will be colored green while
   the thread indent will be colored white.
 - It will look worse for people with tree-style indent unless they
   specifically change the indent option for Slack.
 - The indent will be included when the name is shown in the hotlist,
   and other places that show short_name.

Even without this, it's still possible to get nice indenting of threads
if you add the condition `${buffer.local_variables.slack_type}==thread`
in the indent option and use that to print extra indent for threads. You
will get an extra space before the name (the default indent), but I'll
add a commit to remove that if the indent option includes slack_type.

This reverts commit f55bf93ad9dd9a90afa987ee45ac7c862d85084b.