Remove thread short_name indent if buflist.format.indent checks slack_type
If buflist.format.indent includes "slack_type", assume that it handles
the indent for slack threads specifically, so we should not add any
indent in the short_name. Otherwise, add a space at the start of
short_name so it is indented in the buflist.

This means that you can use e.g. this value for buflist.format.indent to
get a nice tree view in the buflist, with threads nested under the
channels:

    /set buflist.format.indent "${color:237}${if:${buffer.local_variables.slack_type}==thread?${if:${buffer.next_buffer.local_variables.slack_type}==thread?│ ├─:│ └─}:${if:${buffer.next_buffer.local_variables.type}=~^(channel|private)$?├─:└─}}"