Check that user is not None in typing notices
I have experienced this being called with a user id we're not aware of,
in which case metadata["user"] will be None. I think shared
teams/channels are the reason for this.

We should look up the user id when it's unknown, but for now just ignore
typing notices for unknown users so we don't get an exception.