Increase the number of channel members we fetch
conversations.members only returns 100 members if we don't specify
limit, and requires pagination to get the rest. We want to know more
members than that, so specify a limit of 1000, which is the max we can
specify. If a channel has more than 1000 members, we would have to use
papagination to get all, but I'll just do this simple change for now.