Support MPIMs in channel list of rtm.start
For some reason, some MPIMs are listed in the channel list of the
response from rtm.start, with is_mpim set to true, instead of in the
mpims list. These incorrectly appeared as private channels instead of
MPIMs in wee-slack.

These entries do not contain a members list, so we have to call
conversations.members and set the name after receiving the response from
that.