Fix alignment of multiline messages with non-ASCII prefix. (#854)
On Python 2, weechat.string_remove_color doesn't work for Unicode
strings containing non-ASCII characters.  The string must be converted
to bytes using encode_to_utf8.

Furthermore, the length of a non-ASCII string (either in bytes or
codepoints) isn't the same as the width of the string on screen.  To
accurately calculate the width, when the string might contain
combining or double-width characters, use weechat.strlen_screen.

Co-authored-by: Benjamin Moody <benjaminmoody@gmail.com>