Support starting wee-slack without using rtm.start
rtm.start is deprecated and will stop working on September 20, 2022.
This patch replaces it with several other API endpoints to get the info
we need.

The handle_rtmstart method is kept for the test setup to work, but this
setup should also be replaced with the new API endpoints.

This is a necessary step for #699 and #844