Support specifying an auth cookie in slack_api_token
This is necessary for xoxc tokens. The cookie is specified after the
token with a colon separating them, i.e.
"token1,token2:cookiefor2,token3".

This makes slack_api_token even more inconvenient, but it will have to
do until we rewrite the config to have a section for each team, like the
irc config has.

Fixes #844