Don't open all DMs and MPDMs on start
Instead open only unread DMs and MPDMs. Unfortunately it doesn't seem to
be possible to know which DMs or MPDMs are open from the last run or in
the official client (without using internal APIs). We could explore
alternatives like opening the ones with recent messages, or storing
which are open, but this will have to do for now.