Fix message changes not being processed
The Slack API has changed message_changed events from having a unique ts
to having the ts of the message that's changed, which means that the
condition in the start of process_message to not process duplicate
messages (see commit 68966be for info about that) made it skip the
message_changed events.

To fix this, only do this duplication check for messages without a
subtype (i.e. new normal messages). This condition is becoming a bit
convoluted, so it might be better to do this duplication check another
way, but this will have to do for now to get the issue fixed.