Fall back to bot_id as nick if no other info is available
Apparently there are some bots which are not listed in users.list and
which sends messages where the only info about the sender is a bot_id.

To fix this properly we will have to look up the bot with bots.info when
we encounter an unknown bot_id, but for now just fall back to using the
id directly like we do with unknown users to make it possible to use
triggers and filters.