Open open IMs and MPIMs on start
We used the response from rtm.start to find which IMs and MPIMs to open,
but after moving away from that we didn't have the information from the
response from conversations.list, so commit 962bbc5 changed to only open
unread ones.

However, it turns out that conversations.info has the information, and
we are already requesting that for each IM/MPIM to check if it's unread,
so we can use that to open the correct ones.

Fixes #875