Show useful link for Slack posts
The url_private link for Slack posts just downloads a json of the post,
which isn't very useful to click on. The permalink goes to the web view
for a post, which is what we want. The web client adds origin_team and
origin_channel, which makes the post page show info about where the post
is shared, so add that here as well.

We still want to use url_private for images etc., not permalink there as
well, because that links to the Slack client with the image shown in the
sidebar, which is much more inconvenient than just getting the image
directly.