Sleep for 1 ms if 0 ms is specified
asyncio.sleep(0) is used to move the task to the back of the event loop.
This sleep can be used for the same, however if 0 is specified, WeeChat
never calls the callback. This means that the task will silently not
continue which is very bad, so make it sleep for 1 ms instead, which is
practically the same.