Only modify a single chat line for one message in WeeChat 4.0.0
Since WeeChat now has multiline support, the whole message is in one
chat line. This means that we don't need to look for all the lines
belonging to a message and modify them, we can just modify the line for
the message directly. This has the benefit of making it possible to
increase and decrease the number of lines for a message, rather than
having to join extra lines or get blank lines.