Return from modify_buffer_line if the line isn't found
If the line isn't found, line_pointer becomes an empty string which will
crash WeeChat when it's used with hdata_update. This happened when the
history for a channel that contained a reply also posted to the channel
was loaded, because in that case wee-slack tries to modify the parent
message before the channel history is printed, so it couldn't find the
line.