If a message has a user, use that even though it's a bot_message
Messages from Slackbot apparently have both a user and a bot_id, and
sometimes have the subtype bot_message, other times not. However, the
bot_id is B01 which gives bot_not_found when trying to look it up. I see
that the web client shows it with the user profile, and doesn't show the
message as a bot message, so do that here as well to match it and avoid
the invalid bot_id.