Fix links/refs not being processed in normal blocks
In commit 6478c14 the block rendering was changed to escape `<` and `>`.
However, this caused links and refs in normal (not rich_text) blocks to
be shown as refs instead of rendered correctly. The correct behavior is
to not process refs in rich_text blocks, but do it in normal blocks.