Handle millisecond-level timestamps in attachments
The Slack documentation claims that the “ts” flag on attachments
is “An integer Unix timestamp”. However, evidently it can be in
milliseconds (i.e., a timestamp multiplied by 1000), causing
out-of-range exceptions when trying to feed them to date.fromtimestamp().
This can cause all messages after such attachments to just vanish,
as if they were never there, as parsing stops.

Accept both; as a heuristic, the Slack web interface seems to use
100000000000 as a threshold, so we do the same and divide by 1000
if so.