Support mentioning users
This currently has the limitation that the user has to be known to
wee-slack for the mention to work. That is, there needs to be some
interaction with it (message or reaction from or mention) or it has to
be in the loaded members in any of the loaded conversations.