Support completing @nicks in current buffer
Adds all known nicks in the current buffer with an @ prefix to the
nicks completion.

Disables the API search based nick completion for now until it's working
properly. It currently doesn't have any context of which
conversation/thread you are in, so you often get other nicks than the
one you want to complete first. Additionally, it doesn't work with OAuth
tokens.