Add no-secretstorage option for extract_token_from_browser.py
On some Linux systems org.freedesktop.Secret D-Bus service is not
available (installed/enabled), typically as a deliberate choice.  In
those cases `extract_token_from_browser.py` hands for the duration of
D-Bus timeout and eventually fails.  Note that it does not handle the
timeout exception, so a fallback is not used.

This commit adds `--no-secretstorage` option which skips using that
service.