Perform nick completion in last spoken order
This makes it complete nicks that spoke last first, and then nicks that
haven't spoken, and lastly yourself. This is the same as the IRC plugin
does by default.

It adds completion of nicks both with and without @, as the default
completion of nicks from the nicklist is just alphabetical.

Completion of workspace nicks is removed for now. It only completed
workspace nicks that are already loaded, so it might be better to
complete workspace nicks in the `input_complete_cb` function instead
(since functions using `completion_list_add` can't be async, only
`input_complete_cb` can).