Improve performance of rerender_history for WeeChat >= 4.0.0
Previously each message was rerendered individually. This meant that for
each message, we start at the bottom of the buffer and go through all
the lines upwards until we find the message. So the lines were looped
through a lot of times which was slow. Now we instead go through all the
lines once, look up the message for each line and updates it.

This is only for >= 4.0.0 since earlier versions may have multiple lines
for each message (when the message has multiple lines). That would
complicate this way of updating the messages, and I don't think it's
worth implementing it when it's only for an old version.