Configure which files pyright runs on in pyproject.toml
Move the configuration of which files pyright should run on from the CI
command to pyproject.toml.

Pyright is only configured to run on the main files for v3. So don't run
it on the files for the old version of wee-slack, and don't run it on
the util scripts. Ideally we would run it on the util scripts, but they
give a lot of errors now, so we'd have to fix that first.