v2: Fix IM title not being updated on user status change
After we started fetching IM info in check_should_open, updating the
IM title when the user changes their status stopped working because when
info is fetched the members field is updated and the info included your
own user in the members, while the update status code expected only the
other user to be present in members. To fix this, make sure we always
include both users in members and check for that.