Update user info for open IM conversations when using session tokens
When using session tokens we apparently don't get any
user_status_changed events, so when a user updated their status the
title in an IM with them wasn't updated.

If we specify the query param slack_client=desktop when connecting to
the websocket we at least get a user_invalidated event when the user
info changes. This doesn't include any info though, so we have to make a
request to get the updated user info.

On workspaces with a lot of users, we may get user_invalidated very
often, so we can't fetch new user info every time we get it, otherwise
we would make a lot of requests. So at least for now, only update the
user info if we have an open IM with the user. This ensures the status
in the IM title is kept up to date at least.

In other words, nick changes are not picked up and updated. This is also
the case when using OAuth tokens, as we only listen for
user_status_changed events, and not the other user updated events.
Handling nick changes is also a bit more complex, so I'll leave that for
later.