Update read of highlight words after API change
The response from users.prefs.get changed today and doesn't include
highlight_words anymore, which caused wee_slack to crash on startup. I
see that the highlight words are instead available in
`all_notifications_prefs.global.global_keywords` now. Change to read
from this and make it safer so it will print a warning instead of
crashing if this changes again.

Fixes #930