Handle options at start of command and multiple spaces
Previously using an option before the arguments or using multiple spaces
between arguments could cause the command to be interpreted incorrectly.
Specifically, running `/slack reply -alsochannel $123 some text` would
make the first argument be an empty string instead of the message id, so
it would send the reply to the last message in the buffer instead of the
correct one.