Don't default to index 1 in ts_from_hash_or_index
Assuming the last message if hash_or_index is an empty string can easily
lead to mistakes. When the command mentioned in the previous commit
message was interpreted incorrectly, the id ended up as an empty string,
and it got the last message in the buffer. Not finding any message and
getting an error would have been better in that case, so we don't send
replies to the wrong threads.