Don't create new SlackMessages when reloading history
Creating new instances of SlackMessage causes thread_buffer and replies
to be reset. This causes messages that arrive to not be printed in
thread buffers that has been open since before the history was reloaded.

I removed the specific properties for _subscribed and _last_read and
just use the fields in _message_json instead so I don't have to keep
them in sync.