Fix nick completion order when a nick has multiple messages
Previously it would keep the first occurence for each user when finding
the unique senders, but that caused the order to be wrong. Since we want
to complete in the order from the bottom of the history and backwards,
we need to keep the last occurance instead. Do this by reversing twice,
before and after making the list unique.