Prevent repeating tracebacks in Future exceptions
If a Future results in an exception, it is raised every time the future
is awaited. Using a normal raise this caused the traceback to be
appended to the exception every time it happened, which resulted in a
lot of repeated lines in the traceback. To prevent this from happening,
set the traceback when it is raised. This is also what asyncio does.