Include first message when loading history for the whole buffer
When loading history for all the messages that is displayed in the
buffer in the set_hotlist and fill_history methods, the first message
was inadvertently excluded.

This caused a bug where if you had a conversation with only a single
message and last_read was set to 0.0, the conversation would disappear
from the hotlist when reconnecting to the workspace. This happened
because last_read is updated after connecting if it's set to 0.0 and
that causes the buffer to be removed from the hotlist (which is also
something that should be improved). The buffer should be added to the
hotlist again if it has unread messages, but since the first (and in
this case only) message wasn't fetched, that didn't happen.