Handle unread IMs that are missing from client.counts
Apparently if there's an open IM with a deactivated user that DM may be
missing from client.counts, which would result in _initialize_session
raising an error (a user reported this on IRC).

To fix it, change the initialization to merge the info from userBoot and
counts and remove the error thrown if a conversation is missing from
counts. IMs has last_read set from userBoot, so we still have that info
even if it isn't in counts. They also has latest set, however that is
currently always set to the current timestamp for me. So to avoid
incorrect data, delete it, and for the IMs we have in counts set it from
there. The ones missing in counts will then not have latest set, but
we're not currently using this field anyways, we're fetching the history
for the conversation separately.