Add message to _messages in set_hotlist
We never want to recreate messages, but rather update them when we fetch
new info, because they have some internal properties that shouldn't be
reset. So we should store the message in set_hotlist, so it isn't
created again in fill_history. Specifically, the issue was that we set
_last_thread_notify when notifying which may happen from set_hotlist,
but since the message wasn't stored it notified again when fill_history
was run.