Fallback to latest reply for last_read for threads
Only threads that are subscribed to have a last_read property. While
that means we can't store read status across reloads of the script, we
can at least keep a read state for when the script is loaded. By
intializing it to the latest reply, we consider all existing messages in
the thread when the thread parent is loaded to be read. That allows us
to treat any new messages after that as unread.

This fixes an issue with the the newly added
`auto_open_threads_only_subscribed` option that if you disabled it, all
the not subscribed threads would open as unread. Now the threads are
considered read when opened, and new messages in them are considered
unread.