Fix threads not being marked as read after first switch away
WeeChat < 4.5.0 doesn't send the buffer_switch signal if a buffer is
created with buffer_new_props and the display property is set so it
switches automatically to the buffer when it's created. This caused
prev_buffer_pointer to point to the wrong buffer when
signal_buffer_switch_cb was called when the thread buffer was switched
away from, so mark_read wasn't called.

See https://github.com/weechat/weechat/issues/2198