Only set unread and clear hotlist if last_read is the last message
last_read can be set to an older message than the latest one, but this
method currently just moves the unread marker to the bottom and clears
the hotlist, which will be incorrect. So in this case it's better to not
do anything so you don't miss any messages.

A future improvement would be to set the unread marker to the correct
position according to last_read, and set the hotlist to the messages
newer than that. The former is possible with WeeChat >= 4.0.0, and the
latter is possible even with older versions I think.