Unescape HTML entities in block link URLs
Block link URLs don't usually contain HTML entities, but they can e.g.
when created from a bot, or when inserting a link that already contains
a HTML entity in the desktop client (in which case it's removed from the
link in that client).

I see that if a link with an HTML entity is created in a code block in
the desktop client, it shows the text with it escaped, but points to a
link with it unescaped. For code blocks I think it's most important that
the text is correct so since we don't have proper links only show the
escaped text and don't format it with the unescaped link in parenthesis
like we normally would. This is an edge case anyways and probably will
rarely happen.