Only allow options first in some commands
If the command has any arguments that can contain spaces (i.e. all
arguments aren't split on spaces and handled separately), only allow
options to be set before the positional arguments. This is to make it
possible to use words starting with - in the positional arguments and
not have it interpreted as options.