Separate the API methods for command and postMessage
I prefer the methods in slack_api to match one to one with the API
methods since the different API methods take different parameters and
return different responses.

Additionally, we are going to use chat.command for other things in the
future, and then it doesn't make sense to have it under postMessage.
This change also fixes a type error.