Use hook_command_run for the away command
The away command is defined in core and the IRC plugin hooks onto it
with hook_command_run. This changes our command to do the same and makes
it conform to the arguments of the core away command.

If an away message is provided, the presence is set to away and the
status is set to the away message. If no away message is provided, the
presence is set to active and the status is cleared.