Support using the nick color as the user mention color
This adds a new option `slack.color.user_mention_nick_color` which if
set to true will remove the `slack.color.user_mention` option and make
user mentions use the nick color.

Having the enabled property as an init variable in WeeChatOption turned
out to be difficult with dataclasses and getting the typing correct.
Since I don't have any need for creating an option with enabled
initially set to false, just always set it to true when the
WeeChatOption class is instanciated.

Closes #932